requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
sqlalchemy==2.0.36
pyodbc==5.2.0
requests==2.32.3
imapclient==3.0.1
msal==1.30.0
pdfplumber==0.11.4
PyMuPDF==1.24.10
python-docx==1.1.2
pytesseract==0.3.13
easyocr==1.7.1
Pillow==12.2.0
spacy==3.7.5
sentence-transformers==3.0.1
scikit-learn==1.5.2
numpy==2.1.1
pandas==2.2.3
ftfy==6.2.0
